started new server directory for app changes
