+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "youtube-recommendations",
+  "version": "1.0.0",
+  "description": "YouTube clone with focus on educational content",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mannyg91/youtube-recommendations.git"
+  },
+  "author": "Get Rec'd",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/mannyg91/youtube-recommendations/issues"
+  },
+  "homepage": "https://github.com/mannyg91/youtube-recommendations#readme"
+}
