+++ production/server(old)/server.js
@@ -1,0 +1,128 @@
+const express = require('express');
+const helmet = require('helmet'); // enhances security in Express
+const morgan = require('morgan'); // logs all HTTP requests
+const cors = require("cors"); //for cross-origin resource sharing
+const { OAuth2Client } = require("google-auth-library");
+const jwt = require("jsonwebtoken");
+require("dotenv/config");
+
+const app = express();
+
+app.use(helmet()); 
+app.use(morgan('tiny')) // 'tiny' condenses logged HTTP requests
+
+// enables CORS
+app.use(
+    cors({
+      origin: ["http://localhost:3000"],
+      methods: "GET,POST,PUT,DELETE,OPTIONS",
+    })
+  );
+
+app.use(express.json())
+
+// app.get('/', (req, res) => {
+//     res.send('hello world');
+// })
+
+let DB = [];
+
+const GOOGLE_CLIENT_ID = process.env.GOOGLE_CLIENT_ID;
+const client = new OAuth2Client(GOOGLE_CLIENT_ID);
+
+async function verifyGoogleToken(token) {
+  try {
+    const ticket = await client.verifyIdToken({
+      idToken: token,
+      audience: GOOGLE_CLIENT_ID,
+    });
+    console.log(token);
+    return { payload: ticket.getPayload() };
+  } catch (error) {
+    return { error: "Invalid user detected. Please try again" };
+  }
+}
+
+app.post("/signup", async (req, res) => {
+  try {
+    // console.log({ verified: verifyGoogleToken(req.body.credential) });
+    if (req.body.credential) {
+      const verificationResponse = await verifyGoogleToken(req.body.credential);
+
+      if (verificationResponse.error) {
+        return res.status(400).json({
+          message: verificationResponse.error,
+        });
+      }
+
+      const profile = verificationResponse?.payload;
+
+      DB.push(profile);
+
+      res.status(201).json({
+        message: "Signup was successful",
+        user: {
+          firstName: profile?.given_name,
+          lastName: profile?.family_name,
+          picture: profile?.picture,
+          email: profile?.email,
+          token: jwt.sign({ email: profile?.email }, "myScret", {
+            expiresIn: "1d",
+          }),
+        },
+      });
+    }
+  } catch (error) {
+    res.status(500).json({
+      message: "An error occured. Registration failed.",
+    });
+  }
+});
+
+app.post("/login", async (req, res) => {
+  try {
+    if (req.body.credential) {
+      const verificationResponse = await verifyGoogleToken(req.body.credential);
+      if (verificationResponse.error) {
+        return res.status(400).json({
+          message: verificationResponse.error,
+        });
+      }
+
+      const profile = verificationResponse?.payload;
+
+      const existsInDB = DB.find((person) => person?.email === profile?.email);
+
+      if (!existsInDB) {
+        return res.status(400).json({
+          message: "You are not registered. Please sign up",
+        });
+      }
+
+      res.status(201).json({
+        message: "Login was successful",
+        user: {
+          firstName: profile?.given_name,
+          lastName: profile?.family_name,
+          picture: profile?.picture,
+          email: profile?.email,
+          token: jwt.sign({ email: profile?.email }, process.env.JWT_SECRET, {
+            expiresIn: "1d",
+          }),
+        },
+      });
+    }
+  } catch (error) {
+    res.status(500).json({
+      message: error?.message || error,
+    });
+  }
+});
+
+
+
+// Express will run on port 5000 and React will run on 3000
+app.listen(5152, () => {
+    console.log("Server started on port 5152");
+})
+
